+++ Cudheart/Testing/Math/Complex/TestComplex.cu
@@ -1,0 +1,8 @@
+#include "TestComplex.cuh"
+
+namespace Cudheart::Testing::Math {
+	void testComplex() {
+		ComplexType::testComplexType();
+		CPP::ComplexMath::testComplexMath();
+	}
+}