+++ Cudheart/Testing/Math/TestMath.cu
@@ -4,5 +4,6 @@ namespace Cudheart::Testing {
 	void testMath() {
 		Math::testBaseMath();
 		Math::testBitwise();
+		Math::testComplex();
 	}
 }