+++ Changelog.md
@@ -53,6 +53,8 @@
 - removed trueDivide
 - fixed negative
 - base math is now a memeber of the math module lol
+- removed conjugate
+- added angle overload
 
 ## notes:
 -